requirements.txt
flask
tensorflow
pandas
pillow
scikit-learn
joblib
numpy